diffuse material & shadow acne removal

+++ src/color.rs
@@ -10,9 +10,9 @@ pub fn write_color<W: Write>(out: &mut W, pixel_color: &Color) -> () {
     let b = pixel_color.z();
 
     let intensity  = Interval::new(0.000, 0.999);
-    let ir = (256.0 * intensity.clamp(r)) as u32;
-    let ig = (256.0 * intensity.clamp(g)) as u32;
-    let ib = (256.0 * intensity.clamp(b)) as u32;
+    let ir = (256.0 * intensity.clamp(r)) as u8;
+    let ig = (256.0 * intensity.clamp(g)) as u8;
+    let ib = (256.0 * intensity.clamp(b)) as u8;
 
     writeln!(out, "{} {} {}", ir, ig, ib).unwrap_or(())
 }
